refactor(frontend): 結果画面・紙吹雪演出をコンポーネントへ切り出す（issue #607）
設計:
- 選定内容: App.jsx内の`renderResult`関数（1札読み上げ結果画面）を`components/ResultCard.jsx`へ、
  全札読了時の「おめでとう」演出（紙吹雪生成・セッションサマリー・スコア集計・再挑戦ボタン）を
  `components/QuizCompletionScreen.jsx`へそれぞれ切り出した。紙吹雪の疑似乱数生成ロジックも
  QuizCompletionScreen内へ移動し、App.jsx側のuseMemoを削除した。
- 却下内容: 2つを1つのコンポーネントにまとめる案は却下。displayContent.type==='result'
  （1札ごとの結果）とisAllRead（全札読了時の演出）は表示タイミング・依存するpropsが
  異なる別々の画面状態のため、責務ごとに分けた方が見通しがよいと判断した。
- 理由: issue #607（App.jsx 1,820行・App.test.jsx 3,245行の肥大化解消）で提示された
  段階的な対応方針のうち、残っていた「結果画面・演出（renderResult、confetti）→
  コンポーネント化」に対応するため。PR #622（クイズ大会モード管理者側ロジックの
  hook化）に続く2件目の抽出。

影響:
- 影響モジュール: frontend/src/App.jsx（renderResult関数・confettiPieces useMemoを削除し、
  ResultCard/QuizCompletionScreenの呼び出しに置き換え）、frontend/src/components/ResultCard.jsx
  （新規）、frontend/src/components/QuizCompletionScreen.jsx（新規）
- 振る舞いの変更: なし。既存のJSX構造・className・propsの参照元をそのまま移設したのみで、
  画面表示・スタイル・演出タイミングに変更はない
- App.jsx: 1,823行→1,709行

テスト:
- frontend: `npx eslint .` エラー0件
- frontend: `npx vitest run --no-file-parallelism` 全229件成功
  （既存のApp.test.jsxの結果画面・おめでとう画面・紙吹雪に関する統合テストは、
  DOM出力が変更前と同一のためそのまま成功することを確認した。新規コンポーネントは
  App経由の統合テストで従来どおり検証されるため、既存の`components/`配下の慣例
  （QuizRoomBuzzJudgmentModal.jsxも専用テストファイルを持たない）に合わせ、
  今回も専用テストファイルは追加していない）
- frontend: `npm run build` 成功
- backend: `npx eslint .` エラー0件、`npx vitest run --no-isolate` 全1537件成功
  （本変更はfrontendのみだが、karuta固有ルールにより両パッケージを確認）
